Add an abstraction layer between  controller and service
renamed my endpoint ot add-user
refatored my service name to apply to convention
change id type from float to long

+++ src/main/java/com/example/dao/User.java
@@ -10,14 +10,14 @@ import javax.persistence.Table;
 public class User {
 
     @Id
-    private float id;
+    private long id;
 
     private String name;
 
     public User() {
     }
 
-    public User(float id, String name) {
+    public User(long id, String name) {
         this.id = id;
         this.name = name;
     }
@@ -31,7 +31,7 @@ public class User {
         return id;
     }
 
-    public void setId(float id) {
+    public void setId(long id) {
         this.id = id;
     }
